Tried to fix the sound system (got absolutely cooked)
Some certain sounds like the footstep sounds are over a minute or seconds longer, a suggestion would be to make it far more shorter to that like a second or two for it to actually have a fix on the immersion, thus helping with the overall game and it's user experience. But if this request cannot be resolved due to the lack of time we have which is only two days left, then this fix should be enough to somewhat fix the issue being shown.